refactor message a bit
